SoftmaxLossLayer: softmax + log
1. 效果一般，学习速率超过1会优化失败

+++ test/test_net_05_relu.cpp
@@ -110,7 +110,7 @@ int main()
             .train_net_param(NetworkParameter().layer_params(params))
             .test_net_param(NetworkParameter().layer_params(params));
 
-    Optimize<double> o(optimize_param);
+    Optimize<float> o(optimize_param);
 
     o.run();
